Eviction Probability
In this project, client wanted to know which people are likely to get evicted from their homes. But he also wanted to know the exact probability, not just yes or no.

So I had to apply Naive Bayes method to calculate the probability and classify the customers into two categories with 90% plus accuracy.